backend/requirements.txt
# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
gunicorn==23.0.0
websockets==14.1

# Database (async)
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Redis (async)
redis[hiredis]==5.2.0

# Settings / Config
pydantic==2.10.2
pydantic-settings==2.6.1
python-dotenv==1.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Monitoring / Metrics
prometheus-client==0.21.0
prometheus-fastapi-instrumentator==7.0.0

# Utilities
httpx==0.27.2
email-validator==2.2.0
